fix(cli): surface err.cause on import + non-TUI deploy display paths (#1462)
CDK synth failures previously showed only the generic top-level message
("CDK synth failed: ... Subprocess exited with error 1") on the import and
non-TUI deploy paths, dropping the actionable child stderr that the toolkit
wrapper preserves on err.cause.

Centralize the recursive formatError in src/cli/errors.ts (re-exported from
operations/deploy/preflight.ts for existing importers) and route the non-TUI
deploy (commands/deploy/command.tsx) and import (commands/import/command.ts)
display layers through it, so the synth root cause reaches the user. The
existing TUI deploy path already used formatError and is unchanged.
getErrorMessage contract unchanged (message-only).
